Prep for release: - Bumped version number. - Fixed issue where MemberController would throw if Config/debug.php didn't exist. - Fixed warning in setup.php caused by it attempting to delete a non-existant file. - Added the Membership Team group to the 0.3.2.php update script

+++ dev/setup/lib/setup.php
@@ -148,8 +148,11 @@
 					if(!file_exists($templateFilePath))
 					{
 						$fullFilePath = makeAbsolutePath("$configPath$settingName.php");
-						$this->_logMessage("Deleting settings file at $fullFilePath because we couldn't find a matching template");
-						unlink($fullFilePath);
+						if(file_exists($fullFilePath))
+						{
+							$this->_logMessage("Deleting settings file at $fullFilePath because we couldn't find a matching template");
+							unlink($fullFilePath);
+						}
 						continue;
 					}
 				}
